Push because of the exam

+++ taskmaster/program.py
@@ -4,6 +4,7 @@ import datetime
 import threading
 
 from .signal import signals
+from .log import log
 
 class ProgramProperty:
 	def __init__(self, type, default, required=False):
@@ -21,7 +22,8 @@ class ProgramParse:
 		'autostart': ProgramProperty(bool, True, False),
 		'umask': ProgramProperty(int, -1, False),
 		'stoptime': ProgramProperty(int, 1, False),
-		# 'exitcodes': ProgramProperty(int, [0], False),
+		'exitcodes': ProgramProperty(list, [0], False),
+		'startretries': ProgramProperty(int, 0, False),
 	}
 
 	def parseUni(self, program: dict, prop_name):
@@ -34,6 +36,10 @@ class ProgramParse:
 				return
 		if prop.type is not type(program[prop_name]):
 			raise Exception(f'{prop_name}: found {type(program[prop_name])}, {prop.type} expected')
+		if prop.type is list:
+			for i in program[prop_name]:
+				if type(i) is not int:
+					raise Exception(f'{prop_name}: found {type(i)}, int expected')
 		self.__setattr__(prop_name, program[prop_name])
 
 	def parse(self, program: dict):
@@ -44,36 +50,54 @@ class ProgramParse:
 			self.parseUni(program, prop_name)
 
 class Process(subprocess.Popen):
-	def __init__(self, cmd, env, workingdir, umask, name):
+	def __init__(self, name, options):
+		self.name = name
+		self.options = options
+		self.start = False
+		self.retry = self.options.startretries
+		self.gracefulStop = False
+
+	def myStart(self):
 		null = open('/dev/null', 'r')
+		self.start = True
+		self.gracefulStop = False
 		super().__init__(shlex.split(cmd),
-			env=env,
-			cwd=workingdir,
-			umask=umask,
+			env=self.options.env,
+			cwd=self.options.workingdir,
+			umask=self.options.umask,
 			stdin=null, stdout=null, stderr=null
 		)
 		self.start_time = datetime.datetime.now()
-		self.name = name
 
-	def myWait(self):
+	def myWait(self, exitcodes):
 		returnCode = self.wait()
 		print(returnCode)
-		# if returnCode in self.exitcodes:
-		# 	pass # Success
-		# else:
-		# 	pass # Fail
+		if returnCode in exitcodes:
+			log.Info(f'{self.name} end successfuly with code {returnCode}')
+		else:
+			log.Error(f'{self.name} end badly with code {returnCode}')
+			if not self.gracefulStop and self.retry > 0:
+				log.Info(f'Restart process ({self.name})')
+				self.myStart()
+				self.retry -= 1
 
 	def myStop(self, stopsignal, stoptime):
+		if not self.start:
+			log.Warning('Trying to stop a process wich not been start')
+			return
 		self.send_signal(stopsignal)
 		try:
 			self.wait(stoptime)
 		except subprocess.TimeoutExpired:
 			self.kill()
-			print('Force kill')
+			log.Warning('Hard kill. Graceful stop timeout')
+
 	def get_state(self):
 		return "RUNNING"
 
 	def status(self):
+		if not self.start:
+			return f'{self.name} not start\n'
 		uptime = datetime.datetime.now() - self.start_time
 		hours = uptime.total_seconds() // 3600
 		minutes = (uptime.total_seconds() % 3600) // 60
@@ -88,29 +112,27 @@ class Program(ProgramParse):
 		self.launch()
 
 	def launch(self):
-		if self.autostart:
-			self.start()
-
-	def start(self):
 		for index in range(self.numprocs):
 			name = self.name
 			if index:
 				name += f"_{index}"
-			process = Process(self.cmd,
-				env=self.env,
-				workingdir=self.workingdir,
-				umask=self.umask,
-				name=name
+			process = Process(
+				name=name,
+				options=self,
 			)
-			# print(process.pid)
-			thr = threading.Thread(target=process.myWait)
-			thr.start()
 			self.process.append(process)
+		if self.autostart:
+			self.start()
+
+	def start(self):
+		for process in self.process:
+			process.myStart()
+			thr = threading.Thread(target=process.myWait, args=(self.exitcodes,))
+			thr.start()
 	
 	def stop(self):
 		for process in self.process:
 			process.myStop(signals[self.stopsignal], self.stoptime)
-		self.process = []
 
 	def restart(self):
 		self.stop()
